+++ Cargo.toml
@@ -35,7 +35,6 @@ version = "0.1.0"
 [workspace.dependencies]
 cc_core = { path = "crates/cc_core" }
 cc_gui = { path = "crates/cc_gui" }
-cc_oss = { path = "crates/cc_oss" }
 oss_sdk = { path = "crates/oss_sdk" }
 eframe = { version = "0.21.3", features = ["dark-light", "persistence"] }
 serde = { version = "1", features = ["derive"] }
