+++ crates/cc_core/src/lib.rs
@@ -2,8 +2,6 @@ mod error;
 mod history;
 mod image_cache;
 pub mod log;
-// mod object;
-mod oss;
 pub mod runtime;
 mod session;
 mod setting;
@@ -12,11 +10,8 @@ pub mod util;
 pub use error::CoreError;
 pub use history::MemoryHistory;
 pub use image_cache::{ImageCache, ImageFetcher};
-// pub use object::{OssBucket, OssObject, OssObjectType};
-pub use oss::OssClient;
 pub use session::Session;
 pub use setting::{Setting, ShowType};
-pub use tokio;
 pub use tracing;
 
 const LOG_LEVEL: &str = "debug";
