--- crates/cc_oss/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "cc_oss"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-reqwest = { version = "0.11.13", default-features = false }
-base64 = "0.20"
-chrono = "0.4.20"
-log = "0.4.17"
-quick-xml = "0.27"
-derive_more = "0.99.5"
-bytes = "1.0"
-async-trait = "0.1.53"
-httpdate = "1.0.2"
-hmac = "0.12"
-sha1 = "0.10"
-tracing = { workspace = true }
-bytesize = "1.1.0"
-reqsign = { version = "0.8.3", features = ["reqwest_request"] }
