--- crates/cc_oss/src/lib.rs
@@ -1,12 +1,0 @@
-#[macro_use]
-extern crate derive_more;
-
-pub mod errors;
-pub mod object;
-pub mod oss;
-pub mod prelude;
-pub mod query;
-pub mod service;
-
-mod auth;
-mod utils;
