--- crates/cc_oss/src/prelude.rs
@@ -1,3 +1,0 @@
-pub use super::object::*;
-pub use super::oss::OSS;
-pub use super::service::*;
