--- crates/cc_oss/src/utils.rs
@@ -1,16 +1,0 @@
-use super::errors::Error;
-use reqwest::header::{HeaderMap, HeaderName};
-use std::collections::HashMap;
-
-pub fn to_headers<S>(hashmap: HashMap<S, S>) -> Result<HeaderMap, Error>
-where
-    S: AsRef<str>,
-{
-    let mut headers = HeaderMap::new();
-    for (key, val) in hashmap.iter() {
-        let key = key.as_ref();
-        let val = val.as_ref();
-        headers.insert(HeaderName::from_bytes(key.as_bytes())?, val.parse()?);
-    }
-    Ok(headers)
-}
